Added support for dev-portal.json config
