+++ examples/with-zuplo/package.json
@@ -22,6 +22,7 @@
   "dependencies": {
     "react": "18.3.1",
     "react-dom": "18.3.1",
+    "@mdx-js/react": "3.0.1",
     "zudoku": "workspace:*"
   },
   "devDependencies": {
