+++ packages/config/project.json
@@ -4,15 +4,18 @@
   "targets": {
     "build": {
       "inputs": [
-        "{projectRoot}/**/*.ts",
+        "{workspaceRoot}/zudoku/src/**/*.ts",
         "{projectRoot}/tsconfig.json",
         "{projectRoot}/package.json"
       ],
       "outputs": ["{projectRoot}/dist"],
       "cache": true
     },
-    "build:ci": {
+    "build:types": {
       "dependsOn": ["build"]
+    },
+    "build:ci": {
+      "dependsOn": ["build:types"]
     }
   }
 }
