--- packages/config/src/index.ts
@@ -1,5 +1,0 @@
-export {
-  ConfigWithMeta,
-  loadZudokuConfig,
-  zudokuConfigFiles,
-} from "./config.js";
