+++ packages/config/src/index.txt
@@ -1,0 +1,3 @@
+export { type ConfigWithMeta } from "../../zudoku/src/config/common.js";
+export { tryLoadZudokuConfig } from "../../zudoku/src/config/loader.js";
+export { type CommonConfig } from "../../zudoku/src/config/validators/common.js";
