+++ packages/config/src/vite.d.ts
@@ -1,0 +1,18 @@
+// This file is an augmentation to the built-in ImportMeta interface
+// Thus cannot contain any top-level imports
+// <https://www.typescriptlang.org/docs/handbook/declaration-merging.html#module-augmentation>
+
+interface ImportMetaEnv {
+  [key: string]: any;
+  BASE_URL: string;
+  MODE: string;
+  DEV: boolean;
+  PROD: boolean;
+  SSR: boolean;
+}
+
+interface ImportMeta {
+  url: string;
+
+  readonly env: ImportMetaEnv;
+}
