+++ packages/config/tsconfig.txt
@@ -1,7 +1,6 @@
 {
   "compilerOptions": {
     "composite": false,
-    "rootDir": "src",
     "outDir": "dist",
     "target": "ESNext",
     "lib": ["ESNext", "DOM", "DOM.Iterable", "WebWorker"],
@@ -20,17 +19,7 @@
     "skipDefaultLibCheck": true,
     "isolatedModules": true,
     "jsx": "react-jsx",
-    "types": ["vite/client"],
-    "paths": {
-      "zudoku/openapi-worker": ["./src/lib/plugins/openapi-worker.ts"],
-      "zudoku/components": ["./src/lib/components/index.ts"],
-      "zudoku/ui/*": ["./src/lib/ui/*"],
-      "zudoku/plugins/openapi": ["./src/lib/plugins/openapi/index.ts"],
-      "zudoku/plugins/search-inkeep": [
-        "./src/lib/plugins/search-inkeep/index.tsx"
-      ]
-    }
+    "emitDeclarationOnly": true
   },
-  "exclude": ["node_modules", "dist"],
-  "include": ["src/**/*.ts", "src/**/*.tsx"]
+  "include": ["src", "../zudoku/src"]
 }
