+++ packages/zudoku/src/config/common.ts
@@ -1,0 +1,9 @@
+import { CommonConfig } from "./validators/common.js";
+
+export type ConfigWithMeta<TConfig extends CommonConfig> = TConfig & {
+  __meta: {
+    dependencies: string[];
+    path: string;
+    registerDependency: (...file: string[]) => void;
+  };
+};
