+++ packages/zudoku/src/lib/plugins/markdown/index.tsx
@@ -1,7 +1,7 @@
 import type { Toc } from "@stefanprobst/rehype-extract-toc";
 import type { MDXProps } from "mdx/types.js";
 import { RouteObject } from "react-router-dom";
-import { ZudokuDocsConfig } from "../../../config/validators/validate.js";
+import { ZudokuDocsConfig } from "../../../config/validators/common.js";
 import type { ZudokuPlugin } from "../../core/plugins.js";
 import { DocResolver } from "./resolver.js";
