+++ packages/zudoku/src/vite/build.ts
@@ -45,38 +45,35 @@ export async function runBuild(options: { dir: string }) {
     });
 
     const serverConfigFilename = findOutputPathOfServerConfig(serverResult);
-    if (!serverConfigFilename) {
-      throw new Error("Build failed. No server config found");
-    }
+    if (serverConfigFilename) {
+      try {
+        const writtenFiles = await prerender({
+          html,
+          dir: options.dir,
+          base: viteClientConfig.base,
+          serverConfigFilename,
+        });
 
-    try {
-      const writtenFiles = await prerender({
-        html,
-        dir: options.dir,
-        base: viteClientConfig.base,
-        serverConfigFilename,
-      });
+        if (writtenFiles.includes("index.html")) {
+          return;
+        }
 
-      if (writtenFiles.includes("index.html")) {
-        return;
+        await writeFile(
+          path.join(
+            options.dir,
+            "dist",
+            viteClientConfig.base ?? "",
+            "index.html",
+          ),
+          html,
+          "utf-8",
+        );
+      } catch (e) {
+        // dynamic imports in prerender swallow the stack trace, so we log it here
+        // eslint-disable-next-line no-console
+        console.error(e);
       }
-
-      await writeFile(
-        path.join(
-          options.dir,
-          "dist",
-          viteClientConfig.base ?? "",
-          "index.html",
-        ),
-        html,
-        "utf-8",
-      );
-    } catch (e) {
-      // dynamic imports in prerender swallow the stack trace, so we log it here
-      // eslint-disable-next-line no-console
-      console.error(e);
     }
-
     // Write the build output file
     const config = await loadZudokuConfig(options.dir);
     await writeOutput(options.dir, config);
