+++ packages/zudoku/src/vite/dev-server.ts
@@ -6,7 +6,7 @@ import { createServer as createViteServer, type ViteDevServer } from "vite";
 import { type render as serverRender } from "../app/entry.server.js";
 import { logger } from "../cli/common/logger.js";
 import { printDiagnosticsToConsole } from "../cli/common/output.js";
-import type { ConfigWithMeta } from "../config/loader.js";
+import { LoadedConfig } from "../config/config.js";
 import { createGraphQLServer } from "../lib/oas/graphql/index.js";
 import {
   getAppClientEntryPath,
@@ -19,7 +19,7 @@ import { getDevHtml } from "./html.js";
 
 export class DevServer {
   private server: Server | undefined;
-  private currentConfig: ConfigWithMeta | undefined;
+  private currentConfig: LoadedConfig | undefined;
   private terminator: HttpTerminator | undefined;
 
   constructor(private options: { port: number; dir: string; ssr?: boolean }) {}
