+++ packages/zudoku/src/vite/plugin-config-reload.ts
@@ -1,12 +1,11 @@
 import path from "node:path";
 import { type Plugin } from "vite";
 import { printDiagnosticsToConsole } from "../cli/common/output.js";
-import { type ZudokuPluginOptions } from "../config/config.js";
-import { type ConfigWithMeta } from "../config/loader.js";
+import { LoadedConfig, type ZudokuPluginOptions } from "../config/config.js";
 
 export const createConfigReloadPlugin = (
   initialConfig: ZudokuPluginOptions,
-  onConfigChange?: () => Promise<ConfigWithMeta>,
+  onConfigChange?: () => Promise<LoadedConfig>,
 ): [Plugin, () => ZudokuPluginOptions] => {
   let currentConfig = initialConfig;
   let importDependencies = initialConfig.__meta.dependencies;
