+++ packages/zudoku/src/vite/plugin.ts
@@ -1,7 +1,6 @@
 import react from "@vitejs/plugin-react";
 import { type PluginOption } from "vite";
-import { type ZudokuPluginOptions } from "../config/config.js";
-import { type ConfigWithMeta } from "../config/loader.js";
+import { LoadedConfig, type ZudokuPluginOptions } from "../config/config.js";
 import viteApiKeysPlugin from "./plugin-api-keys.js";
 import viteApiPlugin from "./plugin-api.js";
 import viteAuthPlugin from "./plugin-auth.js";
@@ -19,7 +18,7 @@ import { viteSidebarPlugin } from "./plugin-sidebar.js";
 
 export default function vitePlugin(
   initialConfig: ZudokuPluginOptions,
-  onConfigChange?: () => Promise<ConfigWithMeta>,
+  onConfigChange?: () => Promise<LoadedConfig>,
 ): PluginOption {
   const [configReloadPlugin, getCurrentConfig] = createConfigReloadPlugin(
     initialConfig,
