+++ packages/zudoku/src/vite/sitemap.ts
@@ -2,7 +2,7 @@ import { createWriteStream, existsSync } from "fs";
 import { mkdir } from "fs/promises";
 import path from "path";
 import { SitemapStream } from "sitemap";
-import { ZudokuSiteMapConfig } from "../config/validators/validate.js";
+import { ZudokuSiteMapConfig } from "../config/validators/common.js";
 import { joinPath } from "../lib/util/joinPath.js";
 
 export async function generateSitemap({
